Move NamedCache to core/internal
